Restructure challenges/leaderboard files
